Refactor queue monitoring
Uses the new `LinkedList` in place of `Vector` for active kernel
tracking, which allows mostly-lockless concurrent access to the active
kernel set, often with no or minimal allocations.

Propagates errors from the current or prior kernels in `synchronize` by
default, and adds an `errors::Bool` kwarg to `synchronize` to allow
skipping error checking if desired.

Disables all kernel resource cleanup during `wait`, and instead moves
this fully into the kernel monitor. Also fixes a bug in the kernel
monitor where a dead queue may not have all kernel resources cleaned up.

Removes auto-reset of the queue in TLS, and instead provides
`reset_dead_queue!()` to reset the queue if it's dead.

Properly catches and ignores kernel errors in the queue monitor.

Makes the `queue` field in `ROCQueue` no longer atomic, as we do not
change it.